CP-10668: Show error in claim reward screen (#2771)
